Perftest Updates (#1690)
- Fix for failure because of missing references.
- Updated config names in update_references.sh.

+++ jenkins/run_perftest_PR.sh
@@ -15,11 +15,15 @@ for domain in 128 256; do
   ./build/pyutils/driver.py -v -l $logfile perftest run -s $domain $domain 80 -o $result || { echo 'Running failed'; rm -rf $tmpdir; exit 1; }
 
   # create directory for reports
-  mkdir reports
+  mkdir -p reports
   # find references for same configuration
   reference=./pyutils/perftest/references/${label}_$env/$domain.json
-  # plot comparison of current result with references
-  ./build/pyutils/driver.py -v -l $logfile perftest plot compare -i $reference $result -o reports/reference-comparison-$domain || { echo 'Plotting failed'; rm -rf $tmpdir; exit 1; }
+  if [[ -f $reference ]]; then
+    # plot comparison of current result with references
+    ./build/pyutils/driver.py -v -l $logfile perftest plot compare -i $reference $result -o reports/reference-comparison-$domain || { echo 'Plotting failed'; rm -rf $tmpdir; exit 1; }
+  else
+    echo "WARNING: no reference found for config ${label}_$env, domain size $domain" | tee $logfile
+  fi
   # plot comparison between backends
   ./build/pyutils/driver.py -v -l $logfile perftest plot compare-backends -i $result -o reports/backends-comparison-$domain || { echo 'Plotting failed'; rm -rf $tmpdir; exit 1; }
 done
